support timestamp in sst_dump_tool
Summary:
We introduced a new Comparator for timestamp in user keys. In the sst_dump_tool by default we use BytewiseComparator to read sst files. This change allows us to read comparator_name from table properties in meta data block and use it to read.

Test Plan:
make check
Reviewers:

Subscribers:

Tasks:

Tags:
