+++ db/compaction/compaction_job_test.cc
@@ -1292,7 +1292,8 @@ class CompactionJobTimestampTest : public CompactionJobTestBase {
  public:
   CompactionJobTimestampTest()
       : CompactionJobTestBase(test::PerThreadDBPath("compaction_job_ts_test"),
-                              test::ComparatorWithU64Ts(), test::EncodeInt) {}
+                              rocksdb::ComparatorWithU64Ts(), test::EncodeInt) {
+  }
 };
 
 TEST_F(CompactionJobTimestampTest, GCDisabled) {
