+++ test_util/testutil.h
@@ -106,13 +106,6 @@ class SimpleSuffixReverseComparator : public Comparator {
   virtual void FindShortSuccessor(std::string* /*key*/) const override {}
 };
 
-// Returns a user key comparator that can be used for comparing two uint64_t
-// slices. Instead of comparing slices byte-wise, it compares all the 8 bytes
-// at once. Assumes same endian-ness is used though the database's lifetime.
-// Symantics of comparison would differ from Bytewise comparator in little
-// endian machines.
-extern const Comparator* Uint64Comparator();
-
 class StringSink : public FSWritableFile {
  public:
   std::string contents_;
@@ -793,8 +786,6 @@ class ChanglingCompactionFilterFactory : public CompactionFilterFactory {
 // The factory for the hacky skip list mem table that triggers flush after
 // number of entries exceeds a threshold.
 extern MemTableRepFactory* NewSpecialSkipListFactory(int num_entries_per_flush);
-
-extern const Comparator* ComparatorWithU64Ts();
 
 CompressionType RandomCompressionType(Random* rnd);
